requirements.txt
streamlit
openai
python-dotenv
pdfplumber
python-docx
stripe
pandas
flask
requests
